Django/project/requirements.txt
django
mysqlclient
djangorestframework
requests
